# src/configs/swaggerOutput.json
{
	"swagger": "2.0",
	"info": {
		"version": "1.0.0",
		"title": "Hệ thống API E-commerce",
		"description": "Tài liệu API cho hệ thống thương mại điện tử Node.js"
	},
	"host": "localhost:8888",
	"basePath": "/api",
	"schemes": [
		"http",
		"https"
	],
	"securityDefinitions": {
		"bearerAuth": {
			"type": "apiKey",
			"name": "Authorization",
			"in": "header",
			"description": "Header xác thực JWT sử dụng Bearer scheme. Ví dụ: 'Bearer {token}'"
		}
	},
	"definitions": {
		"Brand": {
			"type": "object",
			"properties": {
				"_id": {
					"type": "string",
					"description": "Định danh duy nhất"
				},
				"name": {
					"type": "string",
					"description": "Tên thương hiệu"
				},
				"slug": {
					"type": "string",
					"description": "Tên thương hiệu thân thiện với URL"
				},
				"logoUrl": {
					"type": "string",
					"description": "URL của logo thương hiệu"
				},
				"description": {
					"type": "string",
					"description": "Mô tả thương hiệu"
				},
				"isDomestic": {
					"type": "boolean",
					"description": "Thương hiệu nội địa hay quốc tế"
				},
				"isActive": {
					"type": "boolean",
					"description": "Thương hiệu đang hoạt động hay đã bị xóa mềm"
				},
				"createdAt": {
					"type": "string",
					"format": "date-time",
					"description": "Thời điểm tạo"
				},
				"updatedAt": {
					"type": "string",
					"format": "date-time",
					"description": "Thời điểm cập nhật cuối cùng"
				}
			},
			"required": [
				"name",
				"slug"
			]
		},
...
	},
	"paths": {
		"/brands": {
			"get": {
				"tags": [
					"Brands"
				],
				"summary": "Lấy danh sách tất cả thương hiệu",
				"description": "Lấy danh sách tất cả thương hiệu với các tùy chọn lọc",
				"parameters": [
					{
						"name": "withProductCount",
						"in": "query",
						"type": "boolean",
						"description": "Thêm số lượng sản phẩm cho mỗi thương hiệu"
					},
					{
						"name": "isDomestic",
						"in": "query",
						"type": "boolean",
						"description": "Lọc theo thương hiệu nội địa hoặc quốc tế"
					},
					{
						"name": "groupByOrigin",
						"in": "query",
						"type": "boolean",
						"description": "Nhóm thương hiệu theo nguồn gốc"
					},
					{
						"name": "popular",
						"in": "query",
						"type": "boolean",
						"description": "Lấy các thương hiệu phổ biến nhất"
					},
					{
						"name": "limit",
						"in": "query",
						"type": "integer",
						"description": "Giới hạn số lượng thương hiệu phổ biến trả về"
					}
				],
				"responses": {
					"200": {
						"description": "Danh sách thương hiệu",
						"schema": {
							"type": "array",
							"items": {
								"$ref": "#/definitions/Brand"
							}
						}
					},
					"500": {
						"description": "Lỗi máy chủ"
					}
				}
			},
			"post": {
				"tags": [
					"Brands"
				],
				"summary": "Tạo thương hiệu mới",
				"description": "Tạo một thương hiệu sản phẩm mới",
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"name": "brand",
						"in": "body",
						"description": "Đối tượng thương hiệu cần tạo",
						"required": true,
						"schema": {
							"type": "object",
							"properties": {
								"name": {
									"type": "string",
									"example": "Samsung"
								},
								"logoUrl": {
									"type": "string",
									"example": "https://example.com/samsung.png"
								},
								"description": {
									"type": "string",
									"example": "South Korean electronics company"
								},
								"isDomestic": {
									"type": "boolean",
									"example": false
								}
							},
							"required": [
								"name"
							]
						}
					}
				],
				"responses": {
					"201": {
						"description": "Tạo thương hiệu thành công",
						"schema": {
							"$ref": "#/definitions/Brand"
						}
					},
					"400": {
						"description": "Dữ liệu không hợp lệ"
					},
					"401": {
						"description": "Không được phép - Yêu cầu xác thực"
					},
					"403": {
						"description": "Không có quyền - Yêu cầu quyền Admin"
					},
					"500": {
						"description": "Lỗi máy chủ"
					}
				}
			}
		},
		"/brands/{id}": {
			"get": {
				"tags": [
					"Brands"
				],
				"summary": "Lấy thông tin một thương hiệu theo ID",
				"description": "Lấy thông tin chi tiết của một thương hiệu dựa trên ID",
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"required": true,
						"type": "string",
						"description": "ID của thương hiệu"
					}
				],
				"responses": {
					"200": {
						"description": "Thông tin thương hiệu",
						"schema": {
							"$ref": "#/definitions/Brand"
						}
					},
					"404": {
						"description": "Không tìm thấy thương hiệu"
					},
					"500": {
						"description": "Lỗi máy chủ"
					}
				}
			},
			"put": {
				"tags": [
					"Brands"
				],
				"summary": "Cập nhật thương hiệu",
				"description": "Cập nhật một thương hiệu sản phẩm đã tồn tại theo ID",
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"required": true,
						"type": "string",
						"description": "ID của thương hiệu"
					},
					{
						"name": "brand",
						"in": "body",
						"description": "Thông tin thương hiệu cần cập nhật",
						"required": true,
						"schema": {
							"type": "object",
							"properties": {
								"name": {
									"type": "string",
									"example": "Samsung Electronics"
								},
								"logoUrl": {
									"type": "string",
									"example": "https://example.com/samsung-new.png"
								},
								"description": {
									"type": "string",
									"example": "Updated description for Samsung"
								},
								"isDomestic": {
									"type": "boolean",
									"example": false
								}
							}
						}
					}
				],
				"responses": {
					"200": {
						"description": "Cập nhật thương hiệu thành công",
						"schema": {
							"$ref": "#/definitions/Brand"
						}
					},
					"400": {
						"description": "Dữ liệu không hợp lệ"
					},
					"401": {
						"description": "Không được phép - Yêu cầu xác thực"
					},
					"403": {
						"description": "Không có quyền - Yêu cầu quyền Admin"
					},
					"404": {
						"description": "Không tìm thấy thương hiệu"
					},
					"500": {
						"description": "Lỗi máy chủ"
					}
				}
			},
			"delete": {
				"tags": [
					"Brands"
				],
				"summary": "Xóa thương hiệu vĩnh viễn",
				"description": "Xóa vĩnh viễn một thương hiệu theo ID (chỉ khả thi nếu không có sản phẩm nào sử dụng thương hiệu này)",
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"required": true,
						"type": "string",
						"description": "ID của thương hiệu"
					}
				],
				"responses": {
					"200": {
						"description": "Xóa thương hiệu thành công"
					},
					"400": {
						"description": "Không thể xóa thương hiệu có sản phẩm liên kết"
					},
					"401": {
						"description": "Không được phép - Yêu cầu xác thực"
					},
					"403": {
						"description": "Không có quyền - Yêu cầu quyền Admin"
					},
					"404": {
						"description": "Không tìm thấy thương hiệu"
					},
					"500": {
						"description": "Lỗi máy chủ"
					}
				}
			}
		},
		"/brands/slug/{slug}": {
			"get": {
				"tags": [
					"Brands"
				],
				"summary": "Lấy thông tin một thương hiệu theo slug",
				"description": "Lấy thông tin thương hiệu dựa vào slug",
				"parameters": [
					{
						"name": "slug",
						"in": "path",
						"required": true,
						"type": "string",
						"description": "Slug của thương hiệu"
					}
				],
				"responses": {
					"200": {
						"description": "Thông tin thương hiệu",
						"schema": {
							"$ref": "#/definitions/Brand"
						}
					},
					"404": {
						"description": "Không tìm thấy thương hiệu"
					},
					"500": {
						"description": "Lỗi máy chủ"
					}
				}
			}
		},
		"/brands/soft-delete/{id}": {
			"delete": {
				"tags": [
					"Brands"
				],
				"summary": "Xóa mềm một thương hiệu",
				"description": "Xóa mềm thương hiệu bằng cách đặt isActive thành false",
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"required": true,
						"type": "string",
						"description": "ID của thương hiệu"
					}
				],
				"responses": {
					"200": {
						"description": "Xóa mềm thương hiệu thành công"
					},
					"401": {
						"description": "Không được phép - Yêu cầu xác thực"
					},
					"403": {
						"description": "Không có quyền - Yêu cầu quyền Admin"
					},
					"404": {
						"description": "Không tìm thấy thương hiệu"
					},
					"500": {
						"description": "Lỗi máy chủ"
					}
				}
			}
		},
		"/brands/restore/{id}": {
			"patch": {
				"tags": [
					"Brands"
				],
				"summary": "Khôi phục một thương hiệu đã xóa mềm",
				"description": "Khôi phục thương hiệu đã xóa mềm bằng cách đặt isActive thành true",
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"required": true,
						"type": "string",
						"description": "ID của thương hiệu"
					}
				],
				"responses": {
					"200": {
						"description": "Khôi phục thương hiệu thành công"
					},
					"401": {
						"description": "Không được phép - Yêu cầu xác thực"
					},
					"403": {
						"description": "Không có quyền - Yêu cầu quyền Admin"
					},
					"404": {
						"description": "Không tìm thấy thương hiệu"
					},
					"500": {
						"description": "Lỗi máy chủ"
					}
				}
			}
		},
...
	}
}